\name{CumulE1-package}
\alias{CumulE1-package}
\alias{CumulE1}
\docType{package}
\title{
\packageTitle{CumulE1}
}
\description{
\packageDescription{CumulE1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CumulE1}
\packageIndices{CumulE1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CumulE1}

Maintainer: \packageMaintainer{CumulE1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
